ShakeelFlashBall use FullBar Effect

+++ source/kylarLEDs/LEDInterface/LEDStrip.cpp
@@ -56,7 +56,9 @@ irgb8_t LEDStrip::setHSV(int index, hsv_t hsv){
     hsv.h = ColorUtil::remapHueLUT[(int)(hsv.h/REMAP_LUT_RES)];//ColorUtil::remapHue(hsv.h);
     //timer->add("remapHue(hsv.h);");
 
-    hsv16_t hsv16 = {hsv.h * HSV_HUE_MAX, hsv.s * HSV_SAT_MAX, hsv.v * HSV_VAL_MAX};
+    hsv16_t hsv16 = {static_cast<uint16_t>(hsv.h * HSV_HUE_MAX),
+                     static_cast<uint8_t>(hsv.s * HSV_SAT_MAX),
+                     static_cast<uint8_t>(hsv.v * HSV_VAL_MAX)};
     rgb8_t rgb8;
     ColorUtil::fast_hsv2rgb_32bit(hsv16.h, hsv16.s, hsv16.v, &rgb8.r, &rgb8.g, &rgb8.b);
     //timer->add("::hsv2rgb(hsv)");
